clean up checkbox
